add tab kho tu vung

+++ app/src/main/java/com/minlish/app/data/model/Models.kt
@@ -13,7 +13,7 @@ data class User(
     val avatarUrl: String? = null,
     val englishLevel: String = "A1",        // A1 / A2 / B1 / B2 / C1 / C2
     val learningGoal: String = "General",   // IELTS / TOEIC / Communication / General
-    val dailyTarget: Int = 10,
+    val dailyTarget: Long = 10,
     val notificationTime: String? = "20:00",
     @get:PropertyName("darkMode") @set:PropertyName("darkMode")
     var darkMode: Boolean = false,
@@ -32,8 +32,8 @@ data class VocabSet(
     val description: String? = null,
     val tags: String = "[]",                // JSON array: ["IELTS","Academic"]
     @get:PropertyName("isFavorite") val isFavorite: Boolean = false,
-    val totalWords: Int = 0,                // Cached count
-    val learnedWords: Int = 0,
+    val totalWords: Long = 0,                // Cached count
+    val learnedWords: Long = 0,
     val createdAt: Long = System.currentTimeMillis(),
     val updatedAt: Long = System.currentTimeMillis()
 )
@@ -67,12 +67,12 @@ data class LearningRecord(
     val wordId: String = "",
     val status: String = "NEW",             // NEW / LEARNING / REVIEW / MASTERED
     val easeFactor: Float = 2.5f,           // SM-2: min 1.3
-    val interval: Int = 0,
-    val repetitions: Int = 0,
+    val interval: Long = 0,
+    val repetitions: Long = 0,
     val nextReviewDate: Long = 0L,
-    val lastGrade: Int = 0,                 // 0=Again, 1=Hard, 2=Good, 3=Easy
-    val totalReviews: Int = 0,
-    val correctReviews: Int = 0,
+    val lastGrade: Long = 0,                 // 0=Again, 1=Hard, 2=Good, 3=Easy
+    val totalReviews: Long = 0,
+    val correctReviews: Long = 0,
     val firstLearnedAt: Long? = null,
     val lastReviewedAt: Long? = null
 )
@@ -88,9 +88,9 @@ data class ReviewSession(
     val mode: String = "FLASHCARD",         // FLASHCARD / MULTIPLE_CHOICE / TYPING / LISTENING
     val sessionType: String = "NEW_WORDS",  // NEW_WORDS / REVIEW_DUE / CUSTOM
     val status: String = "IN_PROGRESS",     // IN_PROGRESS / COMPLETED / PAUSED
-    val totalCards: Int = 0,
-    val completedCards: Int = 0,
-    val correctCount: Int = 0,
+    val totalCards: Long = 0,
+    val completedCards: Long = 0,
+    val correctCount: Long = 0,
     val startedAt: Long = System.currentTimeMillis(),
     val endedAt: Long? = null
 )
@@ -103,11 +103,11 @@ data class DailyStats(
     val id: String = "",
     val userId: String = "",
     val date: Long = 0L,                    // Timestamp 00:00 của ngày
-    val newWordsLearned: Int = 0,
-    val wordsReviewed: Int = 0,
-    val correctAnswers: Int = 0,
-    val totalAnswers: Int = 0,
-    val studyMinutes: Int = 0,
+    val newWordsLearned: Long = 0,
+    val wordsReviewed: Long = 0,
+    val correctAnswers: Long = 0,
+    val totalAnswers: Long = 0,
+    val studyMinutes: Long = 0,
     val goalMet: Boolean = false
 )
 
@@ -117,11 +117,11 @@ data class DailyStats(
  */
 data class Streak(
     val userId: String = "",
-    val currentStreak: Int = 0,
-    val longestStreak: Int = 0,
+    val currentStreak: Long = 0,
+    val longestStreak: Long = 0,
     val lastStudyDate: Long = 0L,
-    val totalDaysStudied: Int = 0,
-    val freezesUsed: Int = 0
+    val totalDaysStudied: Long = 0,
+    val freezesUsed: Long = 0
 )
 
 /**
